Refactoring
- Moved class
